build: Create a separate assembly with embedded data
A new Assembly named `GeoTimeZone.Data` that contains the embedded resources is
created and added in the solution. The `Target Framework` of this assembly is
the `netstandard2` which is the common ancestor of all target frameworks, or at
least the frameworks that assembly `GeoTimeZone` targets.

Assembly `GeoTimeZone` references the new assembly. Using this technique the
"shared" assembly is build only for a single frameworks and therefore the size
of the nuget packages (during development) is minimal.